+++ src/routers/productos.js
@@ -1,8 +1,7 @@
 const { Router } = require("express");
 const redis = require("redis");
 const { Producto } = require("../database/models/productos");
-const { midCrearProducto, midIdProducto } = require("../middlewares/productos");
-const { authAdmin } = require("../middlewares/usuarios");
+const { authAdmin } = require("../middlewares/middlewares");
 
 const client = redis.createClient();
 client.on("error", function (error) {
@@ -10,7 +9,9 @@ client.on("error", function (error) {
 })
 
 function makeProductosRouter() {
+
     const router = Router();
+
     // ver productos
     router.get("/productos", async (req, res) => {
         try {
@@ -31,7 +32,7 @@ function makeProductosRouter() {
     })
 
     // crear productos
-    router.post("/productos", authAdmin, midCrearProducto, async (req, res) => {
+    router.post("/productos", authAdmin, async (req, res) => {
         try {
             const p = await Producto.find();
             const nuevoProducto = new Producto();
@@ -49,30 +50,43 @@ function makeProductosRouter() {
             res.status(404).json(`No se ha podido crear el producto`);
         }
     })
+
     // modificar producto
     router.put("/productos/:idProducto", authAdmin, async (req, res) => {
         try {
             const idProducto = Number(req.params.idProducto);
-            const p = await Producto.findOne({ id: idProducto });
-            p.precio = req.body.precio;
-            p.save();
             const productos = await Producto.find();
-            client.set("productos", JSON.stringify(productos));
-            res.status(200).json(`El producto ${p.nombre} ha sido modificado`);
+            if (idProducto > 0 && idProducto <= productos.length) {
+                const p = await Producto.findOne({ id: idProducto });
+                p.precio = req.body.precio;
+                p.save();
+                const productos = await Producto.find();
+                client.set("productos", JSON.stringify(productos));
+                res.status(200).json(`El producto ${p.nombre} ha sido modificado`);
+            } else {
+                res.status(404).json(`Id del producto no existente`);
+            }
         } catch {
-            res.status(404).json(`No se pudo modificar el producto`);
+            res.status(404).json(`Error al modificar el producto`);
         }
     })
+
     // eliminar producto
-    router.delete("/productos/:idProducto", authAdmin, midIdProducto, async (req, res) => {
+    router.delete("/productos/:idProducto", authAdmin, async (req, res) => {
         try {
             const idProducto = Number(req.params.idProducto);
-            const p = await Producto.deleteOne({ id: idProducto });
-            res.status(200).json(`El producto ha sido eliminado`);
+            const productos = await Producto.find();
+            if (idProducto > 0 && idProducto <= productos.length) {
+                await Producto.deleteOne({ id: idProducto });
+                res.status(200).json(`El producto ha sido eliminado`);
+            } else {
+                res.status(404).json(`Id del producto no existente`);
+            }
         } catch {
             res.status(404).json(`No se ha podido eliminar el producto`);
         }
     })
+    
     return router;
 }
